+++ app/Dtos/WordCategoryDto.php
@@ -15,18 +15,34 @@ class WordCategoryDto {
      */
     protected ?string $description;
 
+    /**
+     * @var string $image
+     */
+    protected ?string $image;
 
-    public function __construct(string $name,  ?string $description)
+    /**
+     * @var string $slug
+     */
+    protected string $slug;
+
+    public function __construct(
+        string $name,
+        ?string $description,
+        ?string $image,
+        string $slug
+    )
     {
         $this->name = $name;
-        $this->description = $description;      
+        $this->description = $description;
+        $this->image = $image;
+        $this->slug = $slug;
     }
 
 
     /**
      * @return string
      */
-    public function getName() : string 
+    public function getName() : string
     {
         return $this->name;
     }
@@ -37,5 +53,21 @@ class WordCategoryDto {
     public function getDescription() : ?string
     {
         return $this->description;
+    }
+
+    /**
+     * @return string
+     */
+    public function getImage(): ?string
+    {
+        return $this->image;
+    }
+
+    /**
+     * @return string
+     */
+    public function getSlug(): string
+    {
+        return $this->slug;
     }
 }