+++ app/Models/WordCategory.php
@@ -19,7 +19,7 @@ class WordCategory extends Model
     // protected $primaryKey = 'id';
     // public $timestamps = false;
     protected $guarded = ['id'];
-     protected $fillable = [];
+    protected $fillable = [];
     // protected $hidden = [];
     // protected $dates = [];
 
@@ -34,6 +34,14 @@ class WordCategory extends Model
     | RELATIONS
     |--------------------------------------------------------------------------
     */
+
+    /**
+     * Returns the words related with this category
+     */
+    public function words()
+    {
+        return $this->belongsToMany(RussianWord::class, 'russian_word_pivot_category', 'word_category_id', 'word_id');
+    }
 
     /*
     |--------------------------------------------------------------------------