+++ database/migrations/2021_11_08_160202_create_word_categories_table.php
@@ -18,6 +18,7 @@ class CreateWordCategoriesTable extends Migration
             $table->string('name');
             $table->string('description')->nullable();
             $table->string('image')->nullable();
+            $table->string('slug')->unique();
             $table->timestamps();
         });
     }