+++ resources/views/lessons/russian.blade.php
@@ -20,6 +20,9 @@
     <section class="learning hidden" id="russian-exercise">
         <div class="container max-width">
             <div class="content">
+                <div class="counter">
+                    <div class="percent"></div>
+                </div>
                 <h3 id="demo-text-tajik">Hoops something wrong</h3>
                 <div class="demo-real" id="demo-real">
                     <div class="words" id="choice-demo-words">
@@ -48,7 +51,7 @@
         </div>
         <div class="showAllAnswer hidden">
             <div class="box">
-                <h3 class="center">Ҷавобҳои дурст!</h3>
+                <h3 class="center white">Ҷавобҳои дурст!</h3>
                 <div class="answers-demo p-10" id="allAnswer">
                     {{--   Load here from js --}}
                 </div>
@@ -57,7 +60,7 @@
         </div>
     </section>
 
-    
+
     {{-- question section --}}
     <x-questions-section :questions="$questions"></x-questions-section>