+++ resources/views/words/russian.blade.php
@@ -1,3 +1,0 @@
-@extends('layout.app')
-
-{{ dd($word->words) }}